tests/ci/requirements.txt
cxc-toolkit>=1.0.0
requests==2.32.4
PyYAML==6.0.2